+++ cudax/include/cuda/experimental/__async/sender/when_all.cuh
@@ -491,7 +491,7 @@ _CUDAX_API auto when_all_t::operator()(_Sndrs... __sndrs) const -> __sndr_t<_Snd
 }
 
 template <class... _Sndrs>
-inline constexpr int structured_binding_size<when_all_t::__sndr_t<_Sndrs...>> = sizeof...(_Sndrs) + 2;
+inline constexpr size_t structured_binding_size<when_all_t::__sndr_t<_Sndrs...>> = sizeof...(_Sndrs) + 2;
 
 _CCCL_GLOBAL_CONSTANT when_all_t when_all{};
 } // namespace cuda::experimental::__async
